drman/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
autopep8==1.5.3
black==19.10b0
click==7.1.2
Django==3.0.8
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.9.2
django-filter==2.3.0
pathspec==0.8.0
Pillow==7.2.0
pkg-resources==0.0.0
pycodestyle==2.6.0
pytz==2020.1
regex==2020.6.8
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
